+++ data/src/main/java/com/microsoft/azure/kusto/data/http/HttpClientProperties.java
@@ -13,8 +13,6 @@ public class HttpClientProperties {
     private final boolean keepAlive;
     private final Integer maxKeepAliveTime;
     private final Integer maxConnectionTotal;
-    private final Integer maxConnectionRoute;
-    private final Duration timeout;
     private final Class<? extends HttpClientProvider> provider;
     private final ProxyOptions proxy;
 
@@ -23,8 +21,6 @@ public class HttpClientProperties {
         this.keepAlive = builder.keepAlive;
         this.maxKeepAliveTime = builder.maxKeepAliveTime;
         this.maxConnectionTotal = builder.maxConnectionsTotal;
-        this.maxConnectionRoute = builder.maxConnectionsPerRoute;
-        this.timeout = builder.timeout;
         this.provider = builder.provider;
         this.proxy = builder.proxy;
     }
@@ -58,7 +54,6 @@ public class HttpClientProperties {
      * {@linkplain #maxKeepAliveTime()}.
      *
      * @return whether a custom connection keep-alive strategy should be used
-     *
      * @see #maxKeepAliveTime()
      * @see <a href="https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Keep-Alive">Keep-Alive</a>
      */
@@ -87,24 +82,6 @@ public class HttpClientProperties {
     }
 
     /**
-     * The maximum number of connections the client may keep open at the same time per route.
-     *
-     * @return the maximum number of connections per route
-     */
-    public Integer maxConnectionRoute() {
-        return maxConnectionRoute;
-    }
-
-    /**
-     * The default response timeout to apply to HTTP requests
-     *
-     * @return the timeout
-     */
-    public Duration timeout() {
-        return timeout;
-    }
-
-    /**
      * Gets the HTTP Client Provider used by Azure Core when constructing HTTP Client instances.
      *
      * @return the provider
@@ -128,7 +105,6 @@ public class HttpClientProperties {
         private boolean keepAlive;
         private Integer maxKeepAliveTime = 120;
         private Integer maxConnectionsTotal = 40;
-        private Integer maxConnectionsPerRoute = 40;
         private Duration timeout = Duration.ofMinutes(10);
         private Class<? extends HttpClientProvider> provider = null;
         private ProxyOptions proxy = null;
@@ -161,7 +137,6 @@ public class HttpClientProperties {
          * @param keepAlive set to {@code false} to use a default keep-alive strategy or to {@code true} to use a
          *                  custom one
          * @return the builder instance
-         *
          * @see #maxKeepAliveTime(Integer)
          * @see <a href="https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Keep-Alive">Keep-Alive</a>
          */
@@ -191,17 +166,6 @@ public class HttpClientProperties {
          */
         public HttpClientPropertiesBuilder maxConnectionsTotal(Integer maxConnectionsTotal) {
             this.maxConnectionsTotal = maxConnectionsTotal;
-            return this;
-        }
-
-        /**
-         * Sets the maximum number of connections the client may keep open at the same time for the same route (endpoint).
-         *
-         * @param maxConnections the maximum number of connections per route
-         * @return the builder instance
-         */
-        public HttpClientPropertiesBuilder maxConnectionsPerRoute(Integer maxConnections) {
-            this.maxConnectionsPerRoute = maxConnections;
             return this;
         }