+++ ingest/src/main/java/com/microsoft/azure/kusto/ingest/ManagedStreamingIngestClient.java
@@ -12,6 +12,7 @@ import com.microsoft.azure.kusto.data.auth.ConnectionStringBuilder;
 import com.microsoft.azure.kusto.data.exceptions.DataServiceException;
 import com.microsoft.azure.kusto.data.exceptions.DataWebException;
 import com.microsoft.azure.kusto.data.exceptions.OneApiError;
+import com.microsoft.azure.kusto.data.exceptions.ExceptionsUtils;
 import com.microsoft.azure.kusto.ingest.exceptions.IngestionClientException;
 import com.microsoft.azure.kusto.ingest.exceptions.IngestionServiceException;
 import com.microsoft.azure.kusto.ingest.result.IngestionResult;
@@ -23,10 +24,7 @@ import org.jetbrains.annotations.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.ByteArrayInputStream;
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.io.SequenceInputStream;
+import java.io.*;
 import java.lang.invoke.MethodHandles;
 import java.net.URISyntaxException;
 import java.util.UUID;
@@ -38,19 +36,22 @@ import java.util.UUID;
  * Since the streaming client communicates directly with the engine, it's more prone to failure, so this class
  * holds both a streaming client and a queued client.
  * It tries {@value ATTEMPT_COUNT} times using the streaming client, after which it falls back to the queued streaming client in case of failure.
- * If the size of the stream is bigger than {@value MAX_STREAMING_SIZE_BYTES}, it will fall back to the queued streaming client.
- * <p>
+ * By default the policy for choosing a queued ingestion on the first try is the checking of weather the size of the estimated
+ * raw stream size (a conversion to compressed CSV) is bigger than 4MB, it will fall back to the queued streaming client.
+ * Use {@link #setQueuingPolicy(ManagedStreamingQueuingPolicy)} to override the predicate heuristics.
+ * Use SourceInfo.setRawSizeInBytes to set the raw size of the data. * <p>
  */
 public class ManagedStreamingIngestClient extends IngestClientBase implements QueuedIngestClient {
 
     private static final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
     public static final int ATTEMPT_COUNT = 3;
-    public static final int MAX_STREAMING_SIZE_BYTES = 4 * 1024 * 1024;
     public static final String CLASS_NAME = ManagedStreamingIngestClient.class.getSimpleName();
-    final QueuedIngestClientImpl queuedIngestClient;
+    final QueuedIngestClient queuedIngestClient;
     final StreamingIngestClient streamingIngestClient;
     private final ExponentialRetry exponentialRetryTemplate;
     private HttpClient httpClient = null;
+    private ManagedStreamingQueuingPolicy queuingPolicy = ManagedStreamingQueuingPolicy.Default;
+    private static final String fallbackLogString = "Data size is greater than max streaming size according to the policy. Falling back to queued.";
 
     /**
      * @param dmConnectionString dm connection string
@@ -130,7 +131,7 @@ public class ManagedStreamingIngestClient extends IngestClientBase implements Qu
     /**
      * @param ingestionEndpointConnectionStringBuilder - Endpoint for ingesting data, usually starts with "https://ingest-"
      * @param queryEndpointConnectionStringBuilder     - Endpoint for querying data, does not include "ingest-"
-     * @param autoCorrectEndpoint - Flag to indicate whether to correct the endpoint URI or not
+     * @param autoCorrectEndpoint                      - Flag to indicate whether to correct the endpoint URI or not
      * @throws URISyntaxException if the connection string is invalid
      */
     public ManagedStreamingIngestClient(ConnectionStringBuilder ingestionEndpointConnectionStringBuilder,
@@ -159,8 +160,8 @@ public class ManagedStreamingIngestClient extends IngestClientBase implements Qu
 
     /**
      * @param connectionStringBuilder - Client connection string
-     * @param properties - Additional properties to configure the http client
-     * @param autoCorrectEndpoint - Flag to indicate whether to correct the endpoint URI or not
+     * @param properties              - Additional properties to configure the http client
+     * @param autoCorrectEndpoint     - Flag to indicate whether to correct the endpoint URI or not
      * @throws URISyntaxException if the connection string is invalid
      */
     public ManagedStreamingIngestClient(ConnectionStringBuilder connectionStringBuilder,
@@ -173,8 +174,8 @@ public class ManagedStreamingIngestClient extends IngestClientBase implements Qu
 
     /**
      * @param connectionStringBuilder - Client connection string
-     * @param httpClient - HTTP client
-     * @param autoCorrectEndpoint - Flag to indicate whether to correct the endpoint URI or not
+     * @param httpClient              - HTTP client
+     * @param autoCorrectEndpoint     - Flag to indicate whether to correct the endpoint URI or not
      * @throws URISyntaxException if the connection string is invalid
      */
     public ManagedStreamingIngestClient(ConnectionStringBuilder connectionStringBuilder,
@@ -188,7 +189,7 @@ public class ManagedStreamingIngestClient extends IngestClientBase implements Qu
 
     /**
      * @param ingestionEndpointConnectionStringBuilder - Endpoint for ingesting data, usually starts with "https://ingest-"
-     * @param queryEndpointConnectionStringBuilder - Endpoint for querying data, does not include "ingest-"
+     * @param queryEndpointConnectionStringBuilder     - Endpoint for querying data, does not include "ingest-"
      * @param properties                               - Additional properties to configure the http client
      * @throws URISyntaxException if the connection string is invalid
      */
@@ -246,9 +247,9 @@ public class ManagedStreamingIngestClient extends IngestClientBase implements Qu
 
     /**
      * @param resourceManager ingestion resources manager
-     * @param storageClient - storage utilities
+     * @param storageClient   - storage utilities
      * @param streamingClient - the streaming client
-     * @param retryTemplate - retry template
+     * @param retryTemplate   - retry template
      */
     public ManagedStreamingIngestClient(ResourceManager resourceManager,
             AzureStorageClient storageClient,
@@ -260,6 +261,12 @@ public class ManagedStreamingIngestClient extends IngestClientBase implements Qu
         exponentialRetryTemplate = retryTemplate;
     }
 
+    ManagedStreamingIngestClient(StreamingIngestClient streamingIngestClient, QueuedIngestClient queuedIngestClient, ExponentialRetry exponentialRetry) {
+        this.streamingIngestClient = streamingIngestClient;
+        this.queuedIngestClient = queuedIngestClient;
+        exponentialRetryTemplate = exponentialRetry;
+    }
+
     @Override
     protected IngestionResult ingestFromFileImpl(FileSourceInfo fileSourceInfo, IngestionProperties ingestionProperties)
             throws IngestionClientException, IngestionServiceException {
@@ -269,7 +276,7 @@ public class ManagedStreamingIngestClient extends IngestClientBase implements Qu
         fileSourceInfo.validate();
         ingestionProperties.validate();
         try {
-            StreamSourceInfo streamSourceInfo = IngestionUtils.fileToStream(fileSourceInfo, true);
+            StreamSourceInfo streamSourceInfo = IngestionUtils.fileToStream(fileSourceInfo, true, ingestionProperties.getDataFormat());
             return ingestFromStream(streamSourceInfo, ingestionProperties);
         } catch (FileNotFoundException e) {
             log.error("File not found when ingesting a file.", e);
@@ -292,23 +299,25 @@ public class ManagedStreamingIngestClient extends IngestClientBase implements Qu
 
         BlobClientBuilder blobClientBuilder = new BlobClientBuilder().endpoint(blobSourceInfo.getBlobPath());
         if (httpClient != null) {
-            blobClientBuilder.httpClient((HttpClient) httpClient);
+            blobClientBuilder.httpClient(httpClient);
         }
 
         BlobClient blobClient = blobClientBuilder.buildClient();
+        long blobSize = 0;
         if (blobSourceInfo.getRawSizeInBytes() <= 0) {
             try {
-                blobSourceInfo.setRawSizeInBytes(blobClient.getProperties().getBlobSize());
+                blobSize = blobClient.getProperties().getBlobSize();
             } catch (BlobStorageException e) {
                 throw new IngestionServiceException(
                         blobSourceInfo.getBlobPath(),
-                        "Failed getting blob properties: " + e.getMessage(),
+                        "Failed getting blob properties: " + ExceptionsUtils.getMessageEx(e),
                         e);
             }
         }
 
-        if (blobSourceInfo.getRawSizeInBytes() > MAX_STREAMING_SIZE_BYTES) {
-            log.info("Blob size is greater than max streaming size ({} bytes). Falling back to queued.", blobSourceInfo.getRawSizeInBytes());
+        if (queuingPolicy.shouldUseQueuedIngestion(blobSize, blobSourceInfo.getRawSizeInBytes(),
+                blobSourceInfo.getCompressionType() != null, ingestionProperties.getDataFormat())) {
+            log.info(fallbackLogString);
             return queuedIngestClient.ingestFromBlob(blobSourceInfo, ingestionProperties);
         }
 
@@ -380,7 +389,7 @@ public class ManagedStreamingIngestClient extends IngestClientBase implements Qu
 
     @Override
     protected IngestionResult ingestFromStreamImpl(StreamSourceInfo streamSourceInfo, IngestionProperties ingestionProperties)
-            throws IngestionClientException, IngestionServiceException {
+            throws IngestionClientException, IngestionServiceException, IOException {
         Ensure.argIsNotNull(streamSourceInfo, "streamSourceInfo");
         Ensure.argIsNotNull(ingestionProperties, "ingestionProperties");
 
@@ -392,33 +401,50 @@ public class ManagedStreamingIngestClient extends IngestClientBase implements Qu
             sourceId = UUID.randomUUID();
         }
 
+        streamSourceInfo.setSourceId(sourceId);
         byte[] streamingBytes;
+        InputStream byteArrayStream;
+
+        if (queuingPolicy.shouldUseQueuedIngestion(streamSourceInfo.getStream().available(), streamSourceInfo.getRawSizeInBytes(),
+                streamSourceInfo.getCompressionType() != null, ingestionProperties.getDataFormat())) {
+            log.info(fallbackLogString);
+            return queuedIngestClient.ingestFromStream(streamSourceInfo, ingestionProperties);
+        }
+
         try {
-            streamingBytes = IngestionUtils.readBytesFromInputStream(streamSourceInfo.getStream(), MAX_STREAMING_SIZE_BYTES + 1);
+            if (streamSourceInfo.getStream() instanceof ByteArrayInputStream || streamSourceInfo.getStream() instanceof ResettableFileInputStream) {
+                byteArrayStream = streamSourceInfo.getStream();
+            } else {
+                // If its not a ByteArrayInputStream:
+                // Read 10mb (max streaming size), decide with that if we should stream
+                streamingBytes = IngestionUtils.readBytesFromInputStream(streamSourceInfo.getStream(),
+                        ManagedStreamingQueuingPolicy.MAX_STREAMING_STREAM_SIZE_BYTES + 1);
+                byteArrayStream = new ByteArrayInputStream(streamingBytes);
+                int size = streamingBytes.length;
+                if (queuingPolicy.shouldUseQueuedIngestion(size, streamSourceInfo.getRawSizeInBytes(),
+                        streamSourceInfo.getCompressionType() != null, ingestionProperties.getDataFormat())) {
+                    log.info(fallbackLogString);
+                    StreamSourceInfo managedSourceInfo = new StreamSourceInfo(new SequenceInputStream(byteArrayStream, streamSourceInfo.getStream()),
+                            streamSourceInfo.isLeaveOpen(), sourceId, streamSourceInfo.getCompressionType());
+
+                    return queuedIngestClient.ingestFromStream(managedSourceInfo, ingestionProperties);
+                }
+
+                if (!streamSourceInfo.isLeaveOpen()) {
+                    // From this point we don't need the original stream anymore, we cached it
+                    try {
+                        streamSourceInfo.getStream().close();
+                    } catch (IOException e) {
+                        log.warn("Failed to close stream", e);
+                    }
+                }
+            }
         } catch (IOException e) {
             throw new IngestionClientException("Failed to read from stream.", e);
         }
 
-        // ByteArrayInputStream's close method is a no-op, so we don't need to close it.
-        ByteArrayInputStream byteArrayStream = new ByteArrayInputStream(streamingBytes);
-
-        if (streamingBytes.length > MAX_STREAMING_SIZE_BYTES) {
-            log.info("Stream size is greater than max streaming size ({} bytes). Falling back to queued.", streamingBytes.length);
-            StreamSourceInfo managedSourceInfo = new StreamSourceInfo(new SequenceInputStream(byteArrayStream, streamSourceInfo.getStream()),
-                    streamSourceInfo.isLeaveOpen(), sourceId, streamSourceInfo.getCompressionType());
-            return queuedIngestClient.ingestFromStream(managedSourceInfo, ingestionProperties);
-        }
-
-        if (!streamSourceInfo.isLeaveOpen()) {
-            // From this point we don't need the original stream anymore, we cached it
-            try {
-                streamSourceInfo.getStream().close();
-            } catch (IOException e) {
-                log.warn("Failed to close stream", e);
-            }
-        }
-
-        StreamSourceInfo managedSourceInfo = new StreamSourceInfo(byteArrayStream, true, sourceId, streamSourceInfo.getCompressionType());
+        StreamSourceInfo managedSourceInfo = new StreamSourceInfo(byteArrayStream, true, sourceId, streamSourceInfo.getCompressionType(),
+                streamSourceInfo.getRawSizeInBytes());
         try {
             IngestionResult result = streamWithRetries(managedSourceInfo, ingestionProperties, null);
             if (result != null) {
@@ -433,6 +459,14 @@ public class ManagedStreamingIngestClient extends IngestClientBase implements Qu
                 log.warn("Failed to close byte stream", e);
             }
         }
+    }
+
+    /*
+     * Set the policy that handles the logic over which data size would the client choose to directly use queued ingestion instead of trying streaming ingestion
+     * first.
+     */
+    public void setQueuingPolicy(ManagedStreamingQueuingPolicy queuingPolicy) {
+        this.queuingPolicy = queuingPolicy;
     }
 
     @Override